Remove unnecessary tests and add test for deprecated syntax
